TRIPLEX-810 Добавлено руководство по написанию Stories и примеры для компонента LightBox.
