+++ stories/LightBox/LightBox.stories.tsx
@@ -1,6 +1,14 @@
 import React, { useState } from "react";
 import { Meta, StoryObj } from "@storybook/react";
-import { Controls, Description, Primary, Stories, Heading, Title as SBTitle } from "@storybook/addon-docs/blocks";
+import {
+    Controls,
+    Description,
+    Primary,
+    Stories,
+    Heading,
+    Title as SBTitle,
+    ArgTypes,
+} from "@storybook/addon-docs/blocks";
 import { LightBox } from "../../src/components/LightBox/LightBox";
 import { Page } from "../../src/components/Page/Page";
 import { Button } from "../../src/components/Button/Button";
@@ -16,6 +24,30 @@ import { EIslandType } from "../../src/components/Island/enums";
 import { DefaulticonStrokePrdIcon20 } from "@sberbusiness/icons-next";
 import { FocusTrapUtils } from "../../src/utils/focus/FocusTrapUtils";
 import { EComponentSize } from "../../src/enums/EComponentSize";
+import {
+    DefaultExample,
+    DefaultExampleSource,
+    WithWidePageExample,
+    WithWidePageExampleSource,
+    SmallContentExample,
+    SmallContentExampleSource,
+    SplitModeExample,
+    SplitModeExampleSource,
+    WithSidebarsExample,
+    WithSidebarsExampleSource,
+    WithFixedSidebarsExample,
+    WithFixedSidebarsExampleSource,
+    WithOneSidebarExample,
+    WithOneSidebarExampleSource,
+    WithSideOverlayExample,
+    WithSideOverlayExampleSource,
+    WithSideOverlayLoadingExample,
+    WithSideOverlayLoadingExampleSource,
+    WithTopOverlayExample,
+    WithTopOverlayExampleSource,
+    WithTopOverlayInSideOverlayExample,
+    WithTopOverlayInSideOverlayExampleSource,
+} from "./examples";
 import "./styles.less";
 
 type LightBoxStoryArgs = {
@@ -46,7 +78,7 @@ const meta = {
                     <SBTitle />
                     <Description />
                     <Heading>Props</Heading>
-                    <Controls of={Default} />
+                    <ArgTypes of={LightBox} />
                     <Heading>Playground</Heading>
                     <Primary />
                     <Controls of={Playground} />
@@ -200,7 +232,7 @@ export const Playground: Story = {
             control: { type: "boolean" },
             description: "Показать состояние загрузки LightBox.",
             table: {
-                category: "Состояния",
+                category: "Settings",
                 type: { summary: "boolean" },
                 defaultValue: { summary: "false" },
             },
@@ -209,7 +241,7 @@ export const Playground: Story = {
             control: { type: "boolean" },
             description: "Отображать навигационные кнопки LightBox.Controls.",
             table: {
-                category: "Содержимое",
+                category: "Settings",
                 type: { summary: "boolean" },
                 defaultValue: { summary: "true" },
             },
@@ -218,7 +250,7 @@ export const Playground: Story = {
             control: { type: "boolean" },
             description: "Использовать sticky-позиционирование для Page.Header.",
             table: {
-                category: "Контент",
+                category: "Settings",
                 type: { summary: "boolean" },
                 defaultValue: { summary: "true" },
             },
@@ -227,13 +259,14 @@ export const Playground: Story = {
             control: { type: "boolean" },
             description: "Использовать sticky-позиционирование для Page.Footer.",
             table: {
-                category: "Контент",
+                category: "Settings",
                 type: { summary: "boolean" },
                 defaultValue: { summary: "true" },
             },
         },
     },
     parameters: {
+        testRunner: { skip: true },
         controls: {
             include: ["isLoading", "showControls", "stickyHeader", "stickyFooter"],
         },
@@ -248,432 +281,167 @@ export const Playground: Story = {
     },
 };
 
-export const Default: Story = {
-    render: () => {
-        const [isOpen, setIsOpen] = useState(false);
-
-        const handleOpen = () => setIsOpen(true);
-        const handleClose = () => setIsOpen(false);
-
-        return (
-            <div>
-                <Button theme={EButtonTheme.SECONDARY} size={EComponentSize.MD} onClick={handleOpen}>
-                    Открыть LightBox
-                </Button>
-
-                {isOpen ? (
-                    <LightBox isLoading={false} isSideOverlayOpened={false} isTopOverlayOpened={false}>
-                        <LightBox.Content key="content" isLoading={false}>
-                            <Page>
-                                <Page.Header type={EHeaderPageType.FIRST} sticky>
-                                    <Page.Header.Title>
-                                        <Page.Header.Title.Content>
-                                            <Title
-                                                tag="h1"
-                                                size={ETitleSize.H1}
-                                                tabIndex={-1}
-                                                // Устанавливает фокус на первый элемент при открытии LightBox.
-                                                {...{ [FocusTrapUtils.firstInteractionElementDataAttr]: true }}
-                                            >
-                                                Евгений Онегин
-                                            </Title>
-                                            <Gap size={8} />
-                                            <Text tag="div" size={ETextSize.B3} type={EFontType.SECONDARY}>
-                                                Русский поэт, драматург и прозаик, заложивший основы русского
-                                                реалистического направления.
-                                            </Text>
-                                        </Page.Header.Title.Content>
-                                        <Page.Header.Title.Controls>
-                                            <Button
-                                                icon={<DefaulticonStrokePrdIcon20 paletteIndex={0} />}
-                                                theme={EButtonTheme.SECONDARY}
-                                                size={EComponentSize.MD}
-                                            />
-                                            <Button
-                                                icon={<DefaulticonStrokePrdIcon20 paletteIndex={0} />}
-                                                theme={EButtonTheme.SECONDARY}
-                                                size={EComponentSize.MD}
-                                            />
-                                        </Page.Header.Title.Controls>
-                                    </Page.Header.Title>
-                                </Page.Header>
-
-                                <Page.Body type={EBodyPageType.SECOND}>
-                                    {[0, 1, 2].map((index) => (
-                                        <React.Fragment key={index}>
-                                            <PoemBlock />
-                                            {index < 2 && <Gap size={24} />}
-                                        </React.Fragment>
-                                    ))}
-                                </Page.Body>
-
-                                <Page.Footer type={EFooterPageType.FIRST} sticky>
-                                    <Page.Footer.Description>
-                                        <Page.Footer.Description.Content>А. С. Пушкин</Page.Footer.Description.Content>
-                                        <Page.Footer.Description.Controls>
-                                            <Button theme={EButtonTheme.SECONDARY} size={EComponentSize.MD}>
-                                                Button text
-                                            </Button>
-                                            <Button theme={EButtonTheme.GENERAL} size={EComponentSize.MD}>
-                                                Button text
-                                            </Button>
-                                        </Page.Footer.Description.Controls>
-                                    </Page.Footer.Description>
-                                </Page.Footer>
-                            </Page>
-                        </LightBox.Content>
-
-                        <LightBox.Controls>
-                            <LightBox.Controls.Close
-                                title="Закрыть"
-                                data-test-id="lightbox-close"
-                                onClick={handleClose}
-                            />
-                            <LightBox.Controls.Prev
-                                title="Назад"
-                                clickByArrowLeft
-                                onClick={() => console.log("Prev clicked")}
-                            />
-                            <LightBox.Controls.Next
-                                title="Вперёд"
-                                clickByArrowRight
-                                onClick={() => console.log("Next clicked")}
-                            />
-                        </LightBox.Controls>
-                    </LightBox>
-                ) : null}
-            </div>
-        );
-    },
+export const Default: StoryObj<typeof LightBox> = {
+    name: "Default",
+    render: DefaultExample,
     parameters: {
-        controls: { disable: true },
         docs: {
+            controls: { disable: true },
             description: {
                 story: "Базовая конфигурация LightBox с активными контролами и липкими шапкой и футером.",
+            },
+            source: {
+                code: DefaultExampleSource,
+                language: "tsx",
             },
         },
     },
 };
 
-export const LightBoxWithWidePage: Story = {
-    render: () => {
-        const [isOpen, setIsOpen] = useState(false);
-
-        const handleOpen = () => setIsOpen(true);
-        const handleClose = () => setIsOpen(false);
-
-        return (
-            <div>
-                <Button theme={EButtonTheme.SECONDARY} size={EComponentSize.MD} onClick={handleOpen}>
-                    Открыть LightBox
-                </Button>
-
-                {isOpen ? (
-                    <LightBox isLoading={false} isSideOverlayOpened={false} isTopOverlayOpened={false}>
-                        <LightBox.Content key="content" isLoading={false}>
-                            <Page style={{ maxWidth: "1064px" }}>
-                                <Page.Header type={EHeaderPageType.FIRST} sticky>
-                                    <Page.Header.Title>
-                                        <Page.Header.Title.Content>
-                                            <Title
-                                                tag="h1"
-                                                size={ETitleSize.H1}
-                                                tabIndex={-1}
-                                                // Устанавливает фокус на первый элемент при открытии LightBox.
-                                                {...{ [FocusTrapUtils.firstInteractionElementDataAttr]: true }}
-                                            >
-                                                Евгений Онегин
-                                            </Title>
-                                            <Gap size={8} />
-                                            <Text tag="div" size={ETextSize.B3} type={EFontType.SECONDARY}>
-                                                Русский поэт, драматург и прозаик, заложивший основы русского
-                                                реалистического направления.
-                                            </Text>
-                                        </Page.Header.Title.Content>
-                                        <Page.Header.Title.Controls>
-                                            <Button
-                                                icon={<DefaulticonStrokePrdIcon20 paletteIndex={0} />}
-                                                theme={EButtonTheme.SECONDARY}
-                                                size={EComponentSize.MD}
-                                            />
-                                            <Button
-                                                icon={<DefaulticonStrokePrdIcon20 paletteIndex={0} />}
-                                                theme={EButtonTheme.SECONDARY}
-                                                size={EComponentSize.MD}
-                                            />
-                                        </Page.Header.Title.Controls>
-                                    </Page.Header.Title>
-                                </Page.Header>
-
-                                <Page.Body type={EBodyPageType.SECOND}>
-                                    {[0, 1, 2].map((index) => (
-                                        <React.Fragment key={index}>
-                                            <PoemBlock />
-                                            {index < 2 && <Gap size={24} />}
-                                        </React.Fragment>
-                                    ))}
-                                </Page.Body>
-
-                                <Page.Footer type={EFooterPageType.FIRST} sticky>
-                                    <Page.Footer.Description>
-                                        <Page.Footer.Description.Content>А. С. Пушкин</Page.Footer.Description.Content>
-                                        <Page.Footer.Description.Controls>
-                                            <Button theme={EButtonTheme.SECONDARY} size={EComponentSize.MD}>
-                                                Button text
-                                            </Button>
-                                            <Button theme={EButtonTheme.GENERAL} size={EComponentSize.MD}>
-                                                Button text
-                                            </Button>
-                                        </Page.Footer.Description.Controls>
-                                    </Page.Footer.Description>
-                                </Page.Footer>
-                            </Page>
-                        </LightBox.Content>
-
-                        <LightBox.Controls>
-                            <LightBox.Controls.Close
-                                title="Закрыть"
-                                data-test-id="lightbox-close"
-                                onClick={handleClose}
-                            />
-                            <LightBox.Controls.Prev
-                                title="Назад"
-                                clickByArrowLeft
-                                onClick={() => console.log("Prev clicked")}
-                            />
-                            <LightBox.Controls.Next
-                                title="Вперёд"
-                                clickByArrowRight
-                                onClick={() => console.log("Next clicked")}
-                            />
-                        </LightBox.Controls>
-                    </LightBox>
-                ) : null}
-            </div>
-        );
-    },
+export const LightBoxWithWidePage: StoryObj<typeof LightBox> = {
+    name: "WithWidePage",
+    render: WithWidePageExample,
     parameters: {
-        controls: { disable: true },
         docs: {
+            controls: { disable: true },
             description: {
                 story: "Конфигурация LightBox с Page шириной более 864px. Для этого нужно передать style={{ maxWidth: '1064px' }} в Page.",
             },
-        },
-    },
-};
-
-export const SmallContent: Story = {
-    render: () => {
-        const [isOpen, setIsOpen] = useState(false);
-
-        const handleOpen = () => setIsOpen(true);
-        const handleClose = () => setIsOpen(false);
-
-        return (
-            <div>
-                <Button theme={EButtonTheme.SECONDARY} size={EComponentSize.MD} onClick={handleOpen}>
-                    Открыть LightBox
-                </Button>
-
-                {isOpen ? (
-                    <LightBox isLoading={false} isSideOverlayOpened={false} isTopOverlayOpened={false}>
-                        <LightBox.Content key="content" isLoading={false}>
-                            <Page>
-                                <Page.Header type={EHeaderPageType.FIRST}>
-                                    <Page.Header.Title>
-                                        <Page.Header.Title.Content>
-                                            <Title
-                                                tag="h1"
-                                                size={ETitleSize.H1}
-                                                tabIndex={-1}
-                                                // Устанавливает фокус на первый элемент при открытии LightBox.
-                                                {...{ [FocusTrapUtils.firstInteractionElementDataAttr]: true }}
-                                            >
-                                                Евгений Онегин
-                                            </Title>
-                                            <Gap size={8} />
-                                            <Text tag="div" size={ETextSize.B3} type={EFontType.SECONDARY}>
-                                                Русский поэт, драматург и прозаик, заложивший основы русского
-                                                реалистического направления.
-                                            </Text>
-                                        </Page.Header.Title.Content>
-                                        <Page.Header.Title.Controls>
-                                            <Button
-                                                icon={<DefaulticonStrokePrdIcon20 paletteIndex={0} />}
-                                                theme={EButtonTheme.SECONDARY}
-                                                size={EComponentSize.MD}
-                                            />
-                                            <Button
-                                                icon={<DefaulticonStrokePrdIcon20 paletteIndex={0} />}
-                                                theme={EButtonTheme.SECONDARY}
-                                                size={EComponentSize.MD}
-                                            />
-                                        </Page.Header.Title.Controls>
-                                    </Page.Header.Title>
-                                </Page.Header>
-
-                                <Page.Body type={EBodyPageType.FIRST}>
-                                    Если контента мало, рекомендуется использовать Page.Body типа FIRST, чтобы не было
-                                    пустого пространства.
-                                </Page.Body>
-
-                                <Page.Footer type={EFooterPageType.FIRST}>
-                                    <Page.Footer.Description>
-                                        <Page.Footer.Description.Content>А. С. Пушкин</Page.Footer.Description.Content>
-                                        <Page.Footer.Description.Controls>
-                                            <Button theme={EButtonTheme.SECONDARY} size={EComponentSize.MD}>
-                                                Button text
-                                            </Button>
-                                            <Button theme={EButtonTheme.GENERAL} size={EComponentSize.MD}>
-                                                Button text
-                                            </Button>
-                                        </Page.Footer.Description.Controls>
-                                    </Page.Footer.Description>
-                                </Page.Footer>
-                            </Page>
-                        </LightBox.Content>
-
-                        <LightBox.Controls>
-                            <LightBox.Controls.Close
-                                title="Закрыть"
-                                data-test-id="lightbox-close"
-                                onClick={handleClose}
-                            />
-                            <LightBox.Controls.Prev
-                                title="Назад"
-                                clickByArrowLeft
-                                onClick={() => console.log("Prev clicked")}
-                            />
-                            <LightBox.Controls.Next
-                                title="Вперёд"
-                                clickByArrowRight
-                                onClick={() => console.log("Next clicked")}
-                            />
-                        </LightBox.Controls>
-                    </LightBox>
-                ) : null}
-            </div>
-        );
-    },
-    parameters: {
-        controls: { disable: true },
-        docs: {
-            description: {
-                story: "Базовая конфигурация LightBox с активными контролами и липкими шапкой и футером.",
+            source: {
+                code: WithWidePageExampleSource,
+                language: "tsx",
             },
         },
     },
 };
 
-export const SplitMode: Story = {
-    render: () => {
-        const [isOpen, setIsOpen] = useState(false);
-
-        const handleOpen = () => setIsOpen(true);
-        const handleClose = () => setIsOpen(false);
-
-        return (
-            <div>
-                <Button theme={EButtonTheme.SECONDARY} size={EComponentSize.MD} onClick={handleOpen}>
-                    Открыть LightBox
-                </Button>
-
-                <div id="LightBox-next-view-manager-node-example" />
-
-                {isOpen ? (
-                    <LightBox
-                        isLoading={false}
-                        isSideOverlayOpened={false}
-                        isTopOverlayOpened={false}
-                        lightBoxViewManagerNodeId="LightBox-next-view-manager-node-example"
-                    >
-                        <LightBox.Content key="content" isLoading={false}>
-                            <Page>
-                                <Page.Header type={EHeaderPageType.FIRST} sticky>
-                                    <Page.Header.Title>
-                                        <Page.Header.Title.Content>
-                                            <Title
-                                                tag="h1"
-                                                size={ETitleSize.H1}
-                                                tabIndex={-1}
-                                                // Устанавливает фокус на первый элемент при открытии LightBox.
-                                                {...{ [FocusTrapUtils.firstInteractionElementDataAttr]: true }}
-                                            >
-                                                Евгений Онегин
-                                            </Title>
-                                            <Gap size={8} />
-                                            <Text tag="div" size={ETextSize.B3} type={EFontType.SECONDARY}>
-                                                Русский поэт, драматург и прозаик, заложивший основы русского
-                                                реалистического направления.
-                                            </Text>
-                                        </Page.Header.Title.Content>
-                                        <Page.Header.Title.Controls>
-                                            <Button
-                                                icon={<DefaulticonStrokePrdIcon20 paletteIndex={0} />}
-                                                theme={EButtonTheme.SECONDARY}
-                                                size={EComponentSize.MD}
-                                            />
-                                            <Button
-                                                icon={<DefaulticonStrokePrdIcon20 paletteIndex={0} />}
-                                                theme={EButtonTheme.SECONDARY}
-                                                size={EComponentSize.MD}
-                                            />
-                                        </Page.Header.Title.Controls>
-                                    </Page.Header.Title>
-                                </Page.Header>
-
-                                <Page.Body type={EBodyPageType.SECOND}>
-                                    {[0, 1, 2].map((index) => (
-                                        <React.Fragment key={index}>
-                                            <PoemBlock />
-                                            {index < 2 && <Gap size={24} />}
-                                        </React.Fragment>
-                                    ))}
-                                </Page.Body>
-
-                                <Page.Footer type={EFooterPageType.FIRST} sticky>
-                                    <Page.Footer.Description>
-                                        <Page.Footer.Description.Content>А. С. Пушкин</Page.Footer.Description.Content>
-                                        <Page.Footer.Description.Controls>
-                                            <Button theme={EButtonTheme.SECONDARY} size={EComponentSize.MD}>
-                                                Button text
-                                            </Button>
-                                            <Button theme={EButtonTheme.GENERAL} size={EComponentSize.MD}>
-                                                Button text
-                                            </Button>
-                                        </Page.Footer.Description.Controls>
-                                    </Page.Footer.Description>
-                                </Page.Footer>
-                            </Page>
-                        </LightBox.Content>
-
-                        <LightBox.Controls>
-                            <LightBox.Controls.Close
-                                title="Закрыть"
-                                data-test-id="lightbox-close"
-                                onClick={handleClose}
-                            />
-                            <LightBox.Controls.Prev
-                                title="Назад"
-                                clickByArrowLeft
-                                onClick={() => console.log("Prev clicked")}
-                            />
-                            <LightBox.Controls.Next
-                                title="Вперёд"
-                                clickByArrowRight
-                                onClick={() => console.log("Next clicked")}
-                            />
-                        </LightBox.Controls>
-                    </LightBox>
-                ) : null}
-            </div>
-        );
-    },
+export const SmallContent: StoryObj<typeof LightBox> = {
+    name: "SmallContent",
+    render: SmallContentExample,
     parameters: {
-        controls: { disable: true },
         docs: {
+            controls: { disable: true },
+            description: {
+                story: "Базовая конфигурация LightBox с активными контролами и липкими шапкой и футером.",
+            },
+            source: {
+                code: SmallContentExampleSource,
+                language: "tsx",
+            },
+        },
+    },
+};
+
+export const SplitMode: StoryObj<typeof LightBox> = {
+    name: "SplitMode",
+    render: SplitModeExample,
+    parameters: {
+        docs: {
+            controls: { disable: true },
             description: {
                 story: "Конфигурация LightBox в режиме split-mode. LightBox занимает 90% ширины экрана и располагается по центру. DOM-элемент с идентификатором LightBox-next-view-manager-node-example используется для позиционирования LightBox. Граница этого элемента равна границам LightBox.",
+            },
+            source: {
+                code: SplitModeExampleSource,
+                language: "tsx",
+            },
+        },
+    },
+};
+
+export const WithSidebars: StoryObj<typeof LightBox> = {
+    name: "WithSidebars",
+    render: WithSidebarsExample,
+    parameters: {
+        docs: {
+            controls: { disable: true },
+            source: {
+                code: WithSidebarsExampleSource,
+                language: "tsx",
+            },
+        },
+    },
+};
+
+export const WithFixedSidebars: StoryObj<typeof LightBox> = {
+    name: "WithFixedSidebars",
+    render: WithFixedSidebarsExample,
+    parameters: {
+        docs: {
+            controls: { disable: true },
+            source: {
+                code: WithFixedSidebarsExampleSource,
+                language: "tsx",
+            },
+        },
+    },
+};
+
+export const WithOneSidebar: StoryObj<typeof LightBox> = {
+    name: "WithOneSidebar",
+    render: WithOneSidebarExample,
+    parameters: {
+        docs: {
+            controls: { disable: true },
+            source: {
+                code: WithOneSidebarExampleSource,
+                language: "tsx",
+            },
+        },
+    },
+};
+
+export const WithSideOverlay: StoryObj<typeof LightBox> = {
+    name: "WithSideOverlay",
+    render: WithSideOverlayExample,
+    parameters: {
+        docs: {
+            controls: { disable: true },
+            source: {
+                code: WithSideOverlayExampleSource,
+                language: "tsx",
+            },
+        },
+    },
+};
+
+export const WithSideOverlayLoading: StoryObj<typeof LightBox> = {
+    name: "WithSideOverlay: Loading",
+    render: WithSideOverlayLoadingExample,
+    parameters: {
+        docs: {
+            controls: { disable: true },
+            source: {
+                code: WithSideOverlayLoadingExampleSource,
+                language: "tsx",
+            },
+        },
+    },
+};
+
+export const WithTopOverlay: StoryObj<typeof LightBox> = {
+    name: "WithTopOverlay",
+    render: WithTopOverlayExample,
+    parameters: {
+        docs: {
+            controls: { disable: true },
+            source: {
+                code: WithTopOverlayExampleSource,
+                language: "tsx",
+            },
+        },
+    },
+};
+
+export const WithTopOverlayInSideOverlay: StoryObj<typeof LightBox> = {
+    name: "WithTopOverlayInSideOverlay",
+    render: WithTopOverlayInSideOverlayExample,
+    parameters: {
+        docs: {
+            controls: { disable: true },
+            source: {
+                code: WithTopOverlayInSideOverlayExampleSource,
+                language: "tsx",
             },
         },
     },
